requirements.txt
python-dotenv
requests
tabulate
deezer-python